leetcode/17/letter_combinations_of_a_phone_number_test.go
package main

import (
	"reflect"
	"testing"
)

func Test_letterCombinations(t *testing.T) {
	type args struct {
		digits string
	}
	tests := []struct {
		name string
		args args
		want []string
	}{
		{
			name: "17-1",
			args: args{
				digits: "23",
			},
			want: []string{"ad", "ae", "af", "bd", "be", "bf", "cd", "ce", "cf"},
		},
		{
			name: "17-2",
			args: args{
				digits: "",
			},
			want: []string{},
		},
		{
			name: "17-3",
			args: args{
				digits: "2",
			},
			want: []string{"a", "b", "c"},
		},
		{
			name: "17-4",
			args: args{
				digits: "87",
			},
			want: []string{"tp", "tq", "tr", "ts", "up", "uq", "ur", "us", "vp", "vq", "vr", "vs"},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := letterCombinations(tt.args.digits); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("letterCombinations() = %v, want %v", got, tt.want)
			}
		})
	}
}
